Add 'Scaffold' to class names
