+++ java/test/jmri/util/managers/InternalLightManagerThrowExceptionScaffold.java
@@ -1,53 +1,60 @@
 package jmri.util.managers;
 
+import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import jmri.Sensor;
-import jmri.jmrix.internal.InternalSensorManager;
+import jmri.Light;
+import jmri.jmrix.internal.InternalLightManager;
 
 /**
  * This manager always throws an IllegalArgumentException from provide and get methods.
  * It is used for testing exception handling in tests.
+ * 
+ * The class name ends with 'Scaffold' to exclude it from the coverage statistics,
+ * since it is part of the testing infrastructure.
  */
-public class SensorManagerThrowException extends InternalSensorManager {
+public class InternalLightManagerThrowExceptionScaffold extends InternalLightManager {
 
-    public SensorManagerThrowException() {
-        super("I");
+    public InternalLightManagerThrowExceptionScaffold() {
+        super();
     }
     
     /** {@inheritDoc} */
     @Override
-    protected Sensor createNewSensor(String systemName, String userName) {
+    protected Light createNewLight(String systemName, String userName) {
         throw new IllegalArgumentException("Illegal argument");
     }
     
     /** {@inheritDoc} */
     @Override
-    public Sensor provideSensor(@Nonnull String name) {
+    @Nonnull
+    public Light provideLight(@Nonnull String name) {
         throw new IllegalArgumentException("Illegal argument");
     }
     
     /** {@inheritDoc} */
     @Override
-    public Sensor getSensor(@Nonnull String name) {
+    @CheckForNull
+    public Light getLight(@Nonnull String name) {
         throw new IllegalArgumentException("Illegal argument");
     }
     
     /** {@inheritDoc} */
     @Override
-    public Sensor getBySystemName(@Nonnull String name) {
+    @CheckForNull
+    public Light getBySystemName(@Nonnull String name) {
         throw new IllegalArgumentException("Illegal argument");
     }
     
     /** {@inheritDoc} */
     @Override
-    public Sensor getByUserName(String key) {
+    @CheckForNull
+    public Light getByUserName(String key) {
         throw new IllegalArgumentException("Illegal argument");
     }
     
     /** {@inheritDoc} */
     @Override
-    public Sensor newSensor(@Nonnull String systemName, @Nullable String userName) {
+    public Light newLight(@Nonnull String systemName, @CheckForNull String userName) {
         throw new IllegalArgumentException("Illegal argument");
     }
