+++ java/test/jmri/util/managers/OBlockManagerThrowExceptionScaffold.java
@@ -1,17 +1,19 @@
 package jmri.util.managers;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import jmri.jmrit.logix.OBlock;
 import jmri.jmrit.logix.OBlockManager;
 
 /**
  * This manager always throws an IllegalArgumentException from provide and get methods.
  * It is used for testing exception handling in tests.
+ * 
+ * The class name ends with 'Scaffold' to exclude it from the coverage statistics,
+ * since it is part of the testing infrastructure.
  */
-public class OBlockManagerThrowException extends OBlockManager {
+public class OBlockManagerThrowExceptionScaffold extends OBlockManager {
 
-    public OBlockManagerThrowException() {
+    public OBlockManagerThrowExceptionScaffold() {
         super();
     }
