+++ java/test/jmri/util/managers/WarrantManagerThrowExceptionScaffold.java
@@ -1,54 +1,49 @@
 package jmri.util.managers;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import jmri.Memory;
-import jmri.Turnout;
-import jmri.managers.DefaultMemoryManager;
+import jmri.jmrit.logix.Warrant;
+import jmri.jmrit.logix.WarrantManager;
 
 /**
  * This manager always throws an IllegalArgumentException from provide and get methods.
  * It is used for testing exception handling in tests.
+ * 
+ * The class name ends with 'Scaffold' to exclude it from the coverage statistics,
+ * since it is part of the testing infrastructure.
  */
-public class MemoryManagerThrowException extends DefaultMemoryManager {
+public class WarrantManagerThrowExceptionScaffold extends WarrantManager {
 
-    public MemoryManagerThrowException() {
+    public WarrantManagerThrowExceptionScaffold() {
         super();
     }
     
     /** {@inheritDoc} */
     @Override
-    protected Memory createNewMemory(String systemName, String userName) {
+    public Warrant createNewWarrant(String systemName, String userName, boolean SCWa, long TTP) {
         throw new IllegalArgumentException("Illegal argument");
     }
     
     /** {@inheritDoc} */
     @Override
-    public Memory provideMemory(@Nonnull String name) {
+    public Warrant provideWarrant(@Nonnull String name) {
         throw new IllegalArgumentException("Illegal argument");
     }
     
     /** {@inheritDoc} */
     @Override
-    public Memory getMemory(@Nonnull String name) {
+    public Warrant getWarrant(@Nonnull String name) {
         throw new IllegalArgumentException("Illegal argument");
     }
     
     /** {@inheritDoc} */
     @Override
-    public Memory getBySystemName(@Nonnull String name) {
+    public Warrant getBySystemName(@Nonnull String name) {
         throw new IllegalArgumentException("Illegal argument");
     }
     
     /** {@inheritDoc} */
     @Override
-    public Memory getByUserName(String key) {
-        throw new IllegalArgumentException("Illegal argument");
-    }
-    
-    /** {@inheritDoc} */
-    @Override
-    public Memory newMemory(@Nonnull String systemName, @Nullable String userName) {
+    public Warrant getByUserName(String key) {
         throw new IllegalArgumentException("Illegal argument");
     }
